Extract the `wash find` primaries to a separate package
