+++ cmd/internal/find/primary/boolean.go
@@ -1,4 +1,4 @@
-package find
+package primary
 
 import (
 	"fmt"
