+++ cmd/internal/find/primary/core.go
@@ -1,0 +1,5 @@
+package primary
+
+// LoadAll loads all of `wash find`'s primaries.
+func LoadAll() {
+}
