+++ cmd/internal/find/primary/name_test.go
@@ -1,4 +1,4 @@
-package find
+package primary
 
 import (
 	"testing"
